different fixes

+++ Shamrock/spectrograph.pyx
@@ -116,7 +116,7 @@ class Spectrograph:
         cdef int n = numberpixels
         #cdef array.array values = array.array('f', np.zeros(numberpixels,dtype=np.float))
         #cdef float*  waves[n]
-        cdef np.ndarray[double, ndim=1, mode="c"] waves_numpy = np.empty(numberpixels)
+        cdef np.ndarray[float, ndim=1, mode="c"] waves_numpy = np.empty(numberpixels,dtype=np.float32)
         cdef float *waves = &waves_numpy[0]
         error = lib.ShamrockGetCalibration(self.device, waves, n)
         print(waves_numpy[0])
